refactor: adjust preview route
